Reorganised utils: - merged set ops into set.ts - merged get-instance-path into instance-params - moved formatting utils to lib/formatting/ - moved debug.ts into utils - updated all relevant import paths
