+++ src/app/(protected)/[group]/[subGroup]/[instance]/(admin-panel)/(stage-specific)/(stage-9)/layout.tsx
@@ -3,7 +3,7 @@ import { redirect } from "next/navigation";
 import { Stage } from "@/db/types";
 
 import { api } from "@/lib/trpc/server";
-import { formatParamsAsPath } from "@/lib/utils/general/get-instance-path";
+import { formatParamsAsPath } from "@/lib/utils/general/instance-params";
 import { type InstanceParams } from "@/lib/validations/params";
 
 export default async function Layout({
