+++ src/app/(protected)/[group]/[subGroup]/[instance]/(admin-panel)/edit/_components/wizard-section.tsx
@@ -14,7 +14,7 @@ import {
 import { useInstanceParams } from "@/components/params-context";
 
 import { api } from "@/lib/trpc/client";
-import { formatParamsAsPath } from "@/lib/utils/general/get-instance-path";
+import { formatParamsAsPath } from "@/lib/utils/general/instance-params";
 
 export function WizardSection({
   formDetails,
