+++ src/components/instance-wizard/flag-json-upload/weight-badge.tsx
@@ -1,7 +1,7 @@
 import { Badge } from "@/components/ui/badge";
 
+import { FormatPercent } from "@/lib/formatting/percent";
 import { cn } from "@/lib/utils";
-import { FormatPercent } from "@/lib/utils/format-percent";
 
 export function WeightBadge({
   weight,
