+++ src/lib/utils/clamp.ts
