+++ src/lib/utils/trim.ts
